python/image_adapter/requirements.txt
torch>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
